Removed duplicate test + unnecessary API
