+++ temporal-testing/src/main/java/io/temporal/testing/TestActivityEnvironment.java
@@ -39,7 +39,7 @@ import java.util.Map;
  *   private static class ActivityImpl implements TestActivity {
  *    {@literal @}Override
  *     public String activity1(String input) {
- *       return Activity.getTask().getActivityType().getName() + "-" + input;
+ *       return Activity.getExecutionContext().getInfo().getActivityType().getName() + "-" + input;
  *     }
  *   }
  *
@@ -88,15 +88,11 @@ public interface TestActivityEnvironment {
    * @param <T> Type of the activity interface.
    * @param activityInterface activity interface class that the object under test implements
    * @param options options that specify the activity invocation parameters
-   * @param activityMethodOptions a map keyed on Activity Type Name to its specific invocation
-   *     parameters. By default the name of an activity type is its method name with the first
-   *     letter capitalized.
    * @return The stub that implements the activity interface.
    */
   <T> T newActivityStub(
       Class<T> activityInterface,
-      ActivityOptions options,
-      Map<String, ActivityOptions> activityMethodOptions);
+      ActivityOptions options);
 
   /**
    * Creates a stub that can be used to invoke activities registered through {@link
